feat(theme): run migration when app is started
